authorized_keys: drop CA, use device keys
For the lack of transparency of a personal certificate authority,
I've decided to stop using it in public services. Individual SSH
keys which do not leave their devices are used again. Although
there is no known security problem, buildbot maintainers are
expected to update as soon as possible to purge this CA cert out.